requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
web3==6.11.3
httpx==0.25.2
python-multipart==0.0.6
pydantic==2.5.0
python-dotenv==1.0.0
eth-account==0.10.0
cryptography==41.0.7

# Machine Learning Dependencies
scikit-learn==1.3.2
numpy==1.24.3
pandas==2.0.3
joblib==1.3.2

# Optional: For advanced ML features (uncomment if needed)
# xgboost==2.0.2
# lightgbm==4.1.0